requirements.txt
tflite-runtime
Flask
numpy
pillow
